+++ src/extension.ts
@@ -5,7 +5,7 @@ export function activate(context: vscode.ExtensionContext) {
 	let _statusBarItem: vscode.StatusBarItem;
 	let errorLensEnabled = true;
 
-	let disposableEnableErrorLens = vscode.commands.registerCommand('errorlens.enable', () => {
+	let disposableEnableErrorLens = vscode.commands.registerCommand('errorLens.enable', () => {
 		errorLensEnabled = true;
 
 		const activeTextEditor: vscode.TextEditor | undefined = vscode.window.activeTextEditor;
@@ -16,7 +16,7 @@ export function activate(context: vscode.ExtensionContext) {
 
 	context.subscriptions.push(disposableEnableErrorLens);
 
-	let disposableDisableErrorLens = vscode.commands.registerCommand('errorlens.disable', () => {
+	let disposableDisableErrorLens = vscode.commands.registerCommand('errorLens.disable', () => {
 		errorLensEnabled = false;
 
 		const activeTextEditor: vscode.TextEditor | undefined = vscode.window.activeTextEditor;
@@ -97,6 +97,11 @@ export function activate(context: vscode.ExtensionContext) {
 		const cfg = vscode.workspace.getConfiguration('errorLens');
 		const enabledDiagnosticLevels : string[] = cfg.get('enabledDiagnosticLevels') || ['error', 'warning'];
 		return enabledDiagnosticLevels;
+	}
+	function GetExclude(): Exclude {
+		const cfg = vscode.workspace.getConfiguration('errorLens');
+		const exclude: Exclude = cfg.get('exclude') || [];
+		return exclude;
 	}
 
 	function IsErrorLevelEnabled() {
@@ -250,10 +255,25 @@ export function activate(context: vscode.ExtensionContext) {
 		if (errorLensEnabled) {
 			let aggregatedDiagnostics: IAggregatedDiagnostics = {};
 			let diagnostic: vscode.Diagnostic;
+			const exclude: Exclude = GetExclude();
 
 			// Iterate over each diagnostic that VS Code has reported for this file. For each one, add to
 			// a list of objects, grouping together diagnostics which occur on a single line.
+			nextDiagnostic:
 			for (diagnostic of vscode.languages.getDiagnostics(uriToDecorate)) {
+				// Exclude items specified in `errorLens.exclude` setting
+				for (const excludeItem of exclude) {
+					if (typeof excludeItem === 'string') {
+						if (new RegExp(excludeItem, 'i').test(diagnostic.message)) {
+							continue nextDiagnostic;
+						}
+					} else if (isObject(excludeItem)) {
+						if (diagnostic.source === excludeItem.source &&
+							String(diagnostic.code) === excludeItem.code) {
+							continue nextDiagnostic;
+						}
+					}
+				}
 				const key = `line${diagnostic.range.start.line}`;
 
 				if (aggregatedDiagnostics[key]) {
@@ -455,6 +475,16 @@ export function activate(context: vscode.ExtensionContext) {
 		return str.length > truncationLimit ? str.slice(0, truncationLimit) + '…' : str;
 	}
 }
+
+function isObject(x: any): boolean {
+	return typeof x === 'object' && x !== null;
+}
+
+interface IExcludeObject {
+	code: string;
+	source: string;
+}
+type Exclude = (string | IExcludeObject)[];
 
 // this method is called when your extension is deactivated
 export function deactivate() {